pyproject.toml
[tool.poetry]
name = "acoustid-legacy-search"
version = "0.1.0"
description = ""
authors = ["Lukáš Lalinský <lukas@acoustid.org>"]

[tool.poetry.dependencies]
python = "^3.8"
grpcio = "^1.43.0"
aiohttp = "^3.8.1"
asyncpg = "^0.25.0"
SQLAlchemy = "^1.4.31"

[tool.poetry.dev-dependencies]
grpcio-tools = "^1.43.0"
mypy-protobuf = "^3.2.0"
flake8 = "^4.0.1"
mypy = "^0.931"
grpc-stubs = "^1.24.7"
pytest = "^7.0.0"
sqlalchemy-stubs = "^0.4"

[tool.poetry.scripts]
acoustid-legacy-search-server = "acoustid_legacy_search.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
